refactor: split workflow engine by concept
